+++ lib/honeybadger.ex
@@ -28,7 +28,8 @@ defmodule Honeybadger do
         use_logger: true,
         notice_filter: Honeybadger.NoticeFilter.Default,
         filter: Honeybadger.Filter.Default,
-        filter_keys: [:password, :credit_card]
+        filter_keys: [:password, :credit_card],
+        exclude_errors: []
 
   ### Notifying
 
@@ -287,10 +288,23 @@ defmodule Honeybadger do
       |> contextual_metadata()
       |> Map.put(:breadcrumbs, breadcrumbs)
 
-    exception
-    |> Notice.new(metadata_with_breadcrumbs, stacktrace, fingerprint)
-    |> put_notice_fingerprint()
-    |> Client.send_notice()
+    notice =
+      exception
+      |> Notice.new(metadata_with_breadcrumbs, stacktrace, fingerprint)
+      |> put_notice_fingerprint()
+
+    exclude_error_value = Application.get_env(:honeybadger, :exclude_errors)
+
+    unless exclude_error?(exclude_error_value, notice), do: Client.send_notice(notice)
+  end
+
+  defp exclude_error?(value, notice) when is_list(value) do
+    value = Enum.map(value, &(&1 |> to_string() |> String.trim_leading("Elixir.")))
+    notice.error.class in value
+  end
+
+  defp exclude_error?(value, notice) do
+    value.exclude_error?(notice)
   end
 
   @doc deprecated: "Use Honeybadger.notify/2 instead"